+++ firmware/managed_components/espressif__esp_lcd_ili9341/esp_lcd_ili9341.c
@@ -74,17 +74,9 @@ esp_err_t esp_lcd_new_panel_ili9341(const esp_lcd_panel_io_handle_t io, const es
         break;
     }
 #else
-    switch (panel_dev_config->rgb_ele_order) {
-    case LCD_RGB_ELEMENT_ORDER_RGB:
-        ili9341->madctl_val = 0;
-        break;
-    case LCD_RGB_ELEMENT_ORDER_BGR:
-        ili9341->madctl_val |= LCD_CMD_BGR_BIT;
-        break;
-    default:
-        ESP_GOTO_ON_FALSE(false, ESP_ERR_NOT_SUPPORTED, err, TAG, "unsupported rgb element order");
-        break;
-    }
+    // In ESP-IDF v6.0, rgb_endian is removed from esp_lcd_panel_dev_config_t.
+    // We default to RGB (madctl_val = 0).
+    ili9341->madctl_val = 0;
 #endif
 
     switch (panel_dev_config->bits_per_pixel) {