+++ firmware/managed_components/espressif__adc_mic/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## v0.2.3 - 2026-03-26
+
+### Fixes:
+
+* Suppressed the `'format' is deprecated` warning on some IDF versions.
+
+## v0.2.2 - 2026-03-16
+
+### Fixes:
+
+* Fixed ADC mic sample parsing from ADC DMA output.
+
 ## v0.2.1 - 2025-08-08
 
 ### Fixes: